+++ parser/src/main/java/parser/SyntacticInterface.java
@@ -1,0 +1,7 @@
+package parser;
+
+import common.responses.Result;
+
+public interface SyntacticInterface {
+    Result buildAbstractSyntaxTree();
+}
