+++ parser/src/main/java/parser/ast/builders/ASTreeBuilderInterface.java
@@ -1,0 +1,9 @@
+package parser.ast.builders;
+
+import common.responses.Result;
+import common.tokens.stream.TokenStreamInterface;
+
+public interface ASTreeBuilderInterface {
+    Boolean canBuild(TokenStreamInterface tokenStream);
+    Result build(TokenStreamInterface tokenStream);
+}
