+++ parser/src/main/java/parser/ast/builders/cor/ChainBuilderInterface.java
@@ -1,0 +1,7 @@
+package parser.ast.builders.cor;
+
+import parser.ast.builders.ASTreeBuilderInterface;
+
+public interface ChainBuilderInterface {
+    ASTreeBuilderInterface createDefaultChain();
+}
