+++ parser/src/main/java/parser/factories/AstBuilderFactoryInterface.java
@@ -1,0 +1,14 @@
+package parser.factories;
+
+import parser.ast.builders.ASTreeBuilderInterface;
+
+public interface AstBuilderFactoryInterface {
+    ASTreeBuilderInterface createLetBuilder();
+    ASTreeBuilderInterface createAscriptionBuilder();
+    ASTreeBuilderInterface createIdentifierBuilder();
+    ASTreeBuilderInterface createLiteralBuilder();
+    ASTreeBuilderInterface createTypeBuilder();
+    ASTreeBuilderInterface createBinaryExpressionBuilder();
+    ASTreeBuilderInterface createOperatorBuilder();
+    ASTreeBuilderInterface createFinalBuilder();
+}
