--- src/main/java/syntactic/SyntacticInterface.java
@@ -1,7 +1,0 @@
-package syntactic;
-
-import common.responses.Result;
-
-public interface SyntacticInterface {
-    Result buildAbstractSyntaxTree();
-}
