--- src/main/java/syntactic/ast/builders/ASTreeBuilderInterface.java
@@ -1,9 +1,0 @@
-package syntactic.ast.builders;
-
-import common.responses.Result;
-import common.tokens.stream.TokenStreamInterface;
-
-public interface ASTreeBuilderInterface {
-    Boolean canBuild(TokenStreamInterface tokenStream);
-    Result build(TokenStreamInterface tokenStream);
-}
