--- src/main/java/syntactic/factories/builders/AstBuilderFactoryInterface.java
@@ -1,14 +1,0 @@
-package syntactic.factories.builders;
-
-import syntactic.ast.builders.ASTreeBuilderInterface;
-
-public interface AstBuilderFactoryInterface {
-    ASTreeBuilderInterface createLetBuilder();
-    ASTreeBuilderInterface createAscriptionBuilder();
-    ASTreeBuilderInterface createIdentifierBuilder();
-    ASTreeBuilderInterface createLiteralBuilder();
-    ASTreeBuilderInterface createTypeBuilder();
-    ASTreeBuilderInterface createBinaryExpressionBuilder();
-    ASTreeBuilderInterface createOperatorBuilder();
-    ASTreeBuilderInterface createFinalBuilder();
-}
